リファクタリングの大体完了 ・getMovies.js以外のJSファイルの分割・共通化 ・アクションの共通処理をコンポーネントに切り離し ・アクションの重複処理の共通化
