+++ src/Template/Layout/main.ctp
@@ -58,6 +58,9 @@
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.bundle.min.js" integrity="sha384-1CmrxMRARb6aLqgBO7yyAxTOQE2AKb9GfXnEo760AUcUmFx3ibVJJAzGytlQcNXd" crossorigin="anonymous"></script>
 
   <?php
+  echo $this->Html->script('youtubeapi.js');
+  echo $this->Html->script('getmovies.js');
+  echo $this->Html->script('likedislike.js');
   echo $this->Html->script('main.js');
   ?>
 </body>
